+++ package.json
@@ -22,6 +22,8 @@
         "babel-cli": "^6.26.0",
         "babel-preset-es2015": "^6.24.1",
         "babel-preset-react": "^6.24.1",
-        "nodemon": "^2.0.2"
+        "nodemon": "^2.0.2",
+        "webpack": "^4.41.5",
+        "webpack-cli": "^3.3.10"
     }
 }
