+++ src/App.jsx
@@ -1,3 +1,5 @@
+import IssueAdd from "./IssueAdd.js";
+
 const contentNode = document.getElementById("contents");
 
 const IssueRow = (props) => (
@@ -50,39 +52,6 @@ function IssueTable(props) {
 class IssueFilter extends React.Component {
   render() {
     return <div>Marcador para el filtro de problemas (issue's)</div>;
-  }
-}
-
-class IssueAdd extends React.Component {
-  constructor() {
-    super();
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleSubmit(e) {
-    e.preventDefault();
-    var form = document.forms.issueAdd;
-    this.props.createIssue({
-      status: "New",
-      created: new Date(),
-      owner: form.owner.value,
-      title: form.title.value
-    });
-    // Limpiar el formulario
-    form.owner.value = "";
-    form.title.value = "";
-  }
-
-  render() {
-    return (
-      <div>
-        <form name="issueAdd" onSubmit={this.handleSubmit}>
-          <input type="text" name="owner" placeholder="Propietario" />
-          <input type="text" name="title" placeholder="Título" />
-          <button>Add</button>
-        </form>
-      </div>
-    );
   }
 }
