+++ static/App.js
@@ -2,6 +2,12 @@
 
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
+var _IssueAdd = require("./IssueAdd.js");
+
+var _IssueAdd2 = _interopRequireDefault(_IssueAdd);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
@@ -142,71 +148,20 @@ var IssueFilter = function (_React$Component) {
   return IssueFilter;
 }(React.Component);
 
-var IssueAdd = function (_React$Component2) {
-  _inherits(IssueAdd, _React$Component2);
-
-  function IssueAdd() {
-    _classCallCheck(this, IssueAdd);
-
-    var _this2 = _possibleConstructorReturn(this, (IssueAdd.__proto__ || Object.getPrototypeOf(IssueAdd)).call(this));
-
-    _this2.handleSubmit = _this2.handleSubmit.bind(_this2);
-    return _this2;
-  }
-
-  _createClass(IssueAdd, [{
-    key: "handleSubmit",
-    value: function handleSubmit(e) {
-      e.preventDefault();
-      var form = document.forms.issueAdd;
-      this.props.createIssue({
-        status: "New",
-        created: new Date(),
-        owner: form.owner.value,
-        title: form.title.value
-      });
-      // Limpiar el formulario
-      form.owner.value = "";
-      form.title.value = "";
-    }
-  }, {
-    key: "render",
-    value: function render() {
-      return React.createElement(
-        "div",
-        null,
-        React.createElement(
-          "form",
-          { name: "issueAdd", onSubmit: this.handleSubmit },
-          React.createElement("input", { type: "text", name: "owner", placeholder: "Propietario" }),
-          React.createElement("input", { type: "text", name: "title", placeholder: "T\xEDtulo" }),
-          React.createElement(
-            "button",
-            null,
-            "Add"
-          )
-        )
-      );
-    }
-  }]);
-
-  return IssueAdd;
-}(React.Component);
-
-var IssueList = function (_React$Component3) {
-  _inherits(IssueList, _React$Component3);
+var IssueList = function (_React$Component2) {
+  _inherits(IssueList, _React$Component2);
 
   function IssueList() {
     _classCallCheck(this, IssueList);
 
-    var _this3 = _possibleConstructorReturn(this, (IssueList.__proto__ || Object.getPrototypeOf(IssueList)).call(this));
+    var _this2 = _possibleConstructorReturn(this, (IssueList.__proto__ || Object.getPrototypeOf(IssueList)).call(this));
 
-    _this3.state = {
+    _this2.state = {
       issues: []
     };
 
-    _this3.createIssue = _this3.createIssue.bind(_this3);
-    return _this3;
+    _this2.createIssue = _this2.createIssue.bind(_this2);
+    return _this2;
   }
 
   _createClass(IssueList, [{
@@ -217,7 +172,7 @@ var IssueList = function (_React$Component3) {
   }, {
     key: "loadData",
     value: function loadData() {
-      var _this4 = this;
+      var _this3 = this;
 
       fetch("/api/issues").then(function (response) {
         if (response.ok) {
@@ -229,7 +184,7 @@ var IssueList = function (_React$Component3) {
                 issue.completionDate = new Date(issue.completionDate);
               }
             });
-            _this4.setState({ issues: data.records });
+            _this3.setState({ issues: data.records });
           });
         } else {
           response.json().then(function (error) {
@@ -243,7 +198,7 @@ var IssueList = function (_React$Component3) {
   }, {
     key: "createIssue",
     value: function createIssue(newIssue) {
-      var _this5 = this;
+      var _this4 = this;
 
       fetch("/api/issues", {
         method: "POST",
@@ -257,8 +212,8 @@ var IssueList = function (_React$Component3) {
             if (updateIssue.completionDate) {
               updateIssue.completionDate = new Date(updateIssue.completionDate);
             }
-            var newIssues = _this5.state.issues.concat(updateIssue);
-            _this5.setState({ issues: newIssues });
+            var newIssues = _this4.state.issues.concat(updateIssue);
+            _this4.setState({ issues: newIssues });
           });
         } else {
           response.json().then(function (error) {
@@ -285,7 +240,7 @@ var IssueList = function (_React$Component3) {
         React.createElement("hr", null),
         React.createElement(IssueTable, { issues: this.state.issues }),
         React.createElement("hr", null),
-        React.createElement(IssueAdd, { createIssue: this.createIssue })
+        React.createElement(_IssueAdd2.default, { createIssue: this.createIssue })
       );
     }
   }]);
