+++ static/index.html
@@ -12,7 +12,7 @@
 
 <body>
     <div id="contents"></div>
-    <script src="/App.js"></script>
+    <script src="/app.bundle.js"></script>
 </body>
 
 </html>
